creat stats table and initialize values

+++ app/models/StateObserver.php
@@ -19,8 +19,17 @@ class StateObserver {
   	}
 
 		if  ( $model->state=='ok' ){
-			State::whereState('annotated')->whereHoldingId($model->holding_id)->delete();
-			Note::whereHoldingId($model->holding_id)->delete();
+
+			$model->holding->library->increment('holding_ok')
+
+			// if holding has notes then delete state annotates and notes. Decrement couter of annotated stat
+			if ( State::whereState('annotated')->whereHoldingId($model->holding_id)->exists() ) {
+				
+				State::whereState('annotated')->whereHoldingId($model->holding_id)->delete();
+				Note::whereHoldingId($model->holding_id)->delete();
+				$model->holding->library->decrement('holding_annotated')
+				
+			}
 		}
 
 		if  ( $model->state=='annotated' ){
